requirements.txt
fastapi
uvicorn
python-telegram-bot
sqlalchemy[asyncio]
asyncpg
google-genai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
requests
aiohttp
aiofiles
ffmpeg-python
pillow
pydantic
pydantic-settings
bcrypt
apscheduler
gTTS
numpy
